Move LibraryFile to libraries subdir & namespace
---
