+++ frontend/lib/libraries/CMakeLists.txt
@@ -18,15 +18,6 @@
 target_sources(ChplFrontend-obj
                PRIVATE
 
-               Context.cpp
-               ErrorBase.cpp
-               ErrorMessage.cpp
-               ErrorWriter.cpp
-               ID.cpp
-               Location.cpp
-               TemporaryFileResult.cpp
-               UniqueString.cpp
-               compiler-configuration.cpp
-               serialize-functions.cpp
+               LibraryFile.cpp
 
               )
